+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [0.10] - 2026-01-06
+### Added
+- CI/CD pipeline
+- Profile Patch type
+### Fixed
+- Eliminate 'any' types
+
 ## [0.9] - 2026-01-05
 ### Added
 - Profile create/edit API routes